requirements.txt
flask==3.0.3
flask-cors==4.0.1
gunicorn==21.2.0
openai>=1.54.3,<2.0.0
python-dotenv
pydantic>=2.0.0
PyPDF2==3.0.1
python-docx==1.1.2
psycopg2-binary==2.9.9

# LangChain / LangGraph
langchain-openai>=0.2.8,<0.3.0
langchain-core>=0.3.17,<0.4.0
langchain
langgraph>=0.2.45,<0.3.0

# Azure + OpenAI SDK
azure-core>=1.30.2,<2.0.0
azure-search-documents>=11.6.0b8,<12.0.0
azure-cognitiveservices-speech==1.37.0

# Files
Pillow==10.4.0
pytesseract==0.3.13
requests==2.32.3
